refector : NotificationName Plane 속성으로 변경
