+++ DrawingApp/DrawingApp/Model/Plane.swift
@@ -13,6 +13,15 @@ class Plane{
     private var rectangles: [Rectangle] = []
     var selectedRectangle: Rectangle?
     
+    static let rectangleDidAdd = NSNotification.Name("rectangeDidAdd")
+    static let backgroundDidChagned = NSNotification.Name("backgroundDidAdd")
+    static let alpahDidChanged = NSNotification.Name("alpahDidChanged")
+    static let rectangleDidMutate = NSNotification.Name("rectangleDidMutate")
+    
+    enum UserInfo: String{
+        case updateRectangle
+    }
+    
     func count() -> Int{
         return rectangles.count
     }
@@ -23,7 +32,7 @@ class Plane{
     
     func add(rectangle: Rectangle){
         rectangles.append(rectangle)
-        NotificationCenter.default.post(name: .rectangleDidAdd, object: self, userInfo: [UserInfo.updateRectangle:rectangle])
+        NotificationCenter.default.post(name: Plane.rectangleDidAdd, object: self, userInfo: [UserInfo.updateRectangle:rectangle])
     }
     
     func ExistRectangle(at point: Point) -> Rectangle?{
@@ -40,21 +49,21 @@ class Plane{
     func changeBackgroundColor(to color: Color){
         guard let selected = selectedRectangle else { return }
         selected.changedBackGroundColor(to: color)
-        NotificationCenter.default.post(name: .backgroundDidChagned, object: self, userInfo: [UserInfo.updateRectangle:selected])
+        NotificationCenter.default.post(name: Plane.backgroundDidChagned, object: self, userInfo: [UserInfo.updateRectangle:selected])
          
     }
     
     func changeAlpha(value : Int){
         guard let selected = selectedRectangle else { return }
         selected.changedAlpha(to: value)
-        NotificationCenter.default.post(name: .alpahDidChanged, object: self, userInfo: [UserInfo.updateRectangle:selected])
+        NotificationCenter.default.post(name: Plane.alpahDidChanged, object: self, userInfo: [UserInfo.updateRectangle:selected])
     }
   
     func setSelectedRectangle(_ selected: Rectangle){
         if selected != selectedRectangle {
             self.selectedRectangle = selected
             
-            NotificationCenter.default.post(name: .rectangleDidMutate, object: self, userInfo: [UserInfo.updateRectangle:selected])
+            NotificationCenter.default.post(name: Plane.rectangleDidMutate, object: self, userInfo: [UserInfo.updateRectangle:selected])
         
         }
     }
