--- DrawingApp/DrawingApp/NotificationCenter.swift
@@ -1,19 +1,0 @@
-//
-//  NotificationCenter.swift
-//  DrawingApp
-//
-//  Created by 최예주 on 2022/03/17.
-//
-
-import Foundation
-
-extension Notification.Name{
-    static let rectangleDidAdd = NSNotification.Name("rectangeDidAdd")
-    static let backgroundDidChagned = NSNotification.Name("backgroundDidAdd")
-    static let alpahDidChanged = NSNotification.Name("alpahDidChanged")
-    static let rectangleDidMutate = NSNotification.Name("rectangleDidMutate")
-}
-
-enum UserInfo: String{
-    case updateRectangle
-}
